+++ src/contexts/popup/components/advanced-path-rules-form/handlers/on-delete-button-click.ts
@@ -5,9 +5,9 @@ export const onDeleteButtonClick = (e: Event) => {
     return;
   }
 
-  const { key } = e.currentTarget.dataset;
+  const key = e.currentTarget.dataset['key'] ?? '';
 
-  if (key === undefined || key === '') {
+  if (key === '') {
     return;
   }