+++ src/contexts/popup/components/advanced-path-rules-form/handlers/on-origin-input.ts
@@ -10,9 +10,9 @@ export const onOriginInput = (e: Event) => {
     return;
   }
 
-  const { key } = e.currentTarget.dataset;
+  const key = e.currentTarget.dataset['key'] ?? '';
 
-  if (key === undefined || key === '') {
+  if (key === '') {
     return;
   }